+++ src/bin/swyh-rs.rs
@@ -57,7 +57,7 @@ use swyh_rs::{
 #[global_allocator]
 static GLOBAL: MiMalloc = MiMalloc;
 
-use cpal::traits::StreamTrait;
+use cpal::{SampleFormat, traits::StreamTrait};
 use crossbeam_channel::unbounded;
 use fltk::{app, prelude::ButtonExt};
 use log::{LevelFilter, debug, info};
@@ -205,7 +205,10 @@ fn main() {
     };
 
     // we need to pass some audio config data to the play function
-    let audio_cfg = audio_output_device.default_config();
+    let configured_rate = config.sample_rate.unwrap_or(44100);
+    let audio_cfg = audio_output_device
+        .find_config(configured_rate, SampleFormat::F32, 2)
+        .unwrap_or_else(|| audio_output_device.default_config().clone());
     let wd = WavData {
         sample_format: audio_cfg.sample_format(),
         sample_rate: audio_cfg.sample_rate(),